re-organize doc tracker code
