+++ front/documents_post_process_hooks/hooks/document_tracker/update_tracked_documents/index.ts
@@ -1,0 +1,15 @@
+import { DocumentsPostProcessHook } from "@app/documents_post_process_hooks/hooks";
+import {
+  documentTrackerUpdateTrackedDocumentsOnUpsert,
+  shouldDocumentTrackerUpdateTrackedDocumentsRun,
+} from "@app/documents_post_process_hooks/hooks/document_tracker/update_tracked_documents/lib";
+
+// This hook is meant to update the TrackedDocuments table based on
+// DUST_TRACK tags seen in documents. It should run on upserts, and on deletes (to cleanup)
+export const documentTrackerUpdateTrackedDocumentsPostProcessHook: DocumentsPostProcessHook =
+  {
+    type: "document_tracker_update_tracked_documents",
+    getDebounceMs: async () => 1000,
+    filter: shouldDocumentTrackerUpdateTrackedDocumentsRun,
+    onUpsert: documentTrackerUpdateTrackedDocumentsOnUpsert,
+  };
